+++ src/main/kotlin/ru/citeck/launcher/core/config/bundle/BundleNotFoundException.kt
@@ -1,0 +1,6 @@
+package ru.citeck.launcher.core.config.bundle
+
+class BundleNotFoundException(val ref: BundleRef) :
+    RuntimeException(
+        "Bundle is not found by key '${ref.key}' in repo ${ref.repo}"
+    )