+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release 1.1.3
 
+## New features
+
+* Added the ability to edit a namespace without stopping all services
+* Added the option to update kits from the repository directly in the namespace edit form
+* Renamed page title to **“Citeck Launcher”**
+
 ## Fixes
 
 * Fixed scrolling issue in the app definition editor